Optimize modules release time (#326)
We should run with `-m` option to enable parallel uploads. Additionally,
we should run with either `-c` or `-i` instead of the default (which
compares modification time). Since files are sync'd from a
actions/checkout git clone, all of the files mtimes are set to the time
of checkout, so the default behavior will be to copy every file every
time.