Open a single PR if changes are found
Currently, the fetch.yaml workflow will create a unique PR each time it
detects changes. This means that if we don't merge an update one day,
we'll get duplicate PRs (of which only one can be merged without merge
conflicts on the other).

Update the workflow to use a consistent branch name to avoid this (there
will only ever be a single PR open at a time).