Update protocolbuffers/wellknowntypes module to include go_features.proto (#798)
As of v29.1, the protobuf releases include the
`google/protobuf/go_features.proto` import. This
tiny change will cause that file to be included in the
protocolbuffers/wellknowntypes module.

So, with the next release of protobuf and the wellknowntypes
module (v29.2 or v30.0, whichever comes next), we can stop
sync'ing the protocolbuffers/gofeatures module and update
its README to indicate that it is deprecated and that its
contents are now available from the wellknowntypes
module instead.